requirements.txt
Django==1.4.1
PyYAML==3.10
django-configurations==0.1
sorl-thumbnail==11.12
wsgiref==0.1.2
